Refactor code for exporting spectral library into new class "SpectralLibraryExporter".
